Update ChapterViewController to adopt a viewmodel
